- implemented calculation of view entropy for a single timestep (untested!!!!)

+++ src/util/util.hpp
@@ -83,7 +83,7 @@ namespace util
 
     };
 
-    using bin_t = std::tuple<double, double, unsigned int>;
+    using bin_t = std::tuple<double, double, size_t>;
     //-------------------------------------------------------------------------
     // Templated functions
     //-------------------------------------------------------------------------
